Handle import of django urls correctly.
Issue #264 lead to commit 1e8049d5b, which first attempted to import
from the old path, and on ImportError would import from the new path.
However, doing it that way leads to DeprecationWarnings bubbling up
in Django 1.5. The correct way is to try the new path *first*, and
if that fails, use the old path. This way it is backwards-compatible
and prevents DeprecationWarnings.

(@Troytft had it correct with commit 4b992de26a2 in issue #267.)