Wrap the "grp-navigation" div in admin/base.html in a block.
Sometimes users would like to customize the navigation bar.
To do this, they can use the "nav-global" block or the "userlinks" block,
but this approach has one issue: these blocks are both inside
the "grp-navigation" div, which has both CSS and JS applied to it.
The only way for the user to do what he wants is to override admin/base.html 
(which is a big file, so this approach is prone to errors) or take out the div id
by a script (which is ugly). Wrapping the whole div in a block is a small 
modification and an elegant solution.