Fix #593: RelatedObjectLookups popup window
This fixes the problem of adding a related object to a model in admin.
Moving from Django 1.7 to 1.8, they decided to decouple the html
template from the javascript logic. In 1.7 they had an `onclick` in the
html tag itself, but in 1.8 they moved it out and attached a click
handler that performed the same functionality.

This commit pulls in the JS that adds those click handlers.