Make formset sorting work with Django 1.10
Formset sorting was done by extracting the field to sort by from each formset
item, creating a `[(field, formset_item)]` list and sorting that list. For this
to work `formset_item` (instance of `InlineAdminForm`) had to be comparable. I
don't know about earlier versions, but at least with Django 1.10 it isn't.

Sorting is now done by partitioning the items that can be sorted from the ones
that can't, sorting the ones that can be sorted, and returning the sorted ones
followed by the unsortable ones.